Flesh out grid point optimization formulation
Somehow my original version of the document was deleted.  Fortunately I
had sent a pdf copy of the report to CIRS via email, and was able to
recover most of it from there.

While re-typing in the last section, I had a valuable realization:

Rho, the cutoff for determine whether two points are matching, should
be a function the distance from the isocenter!  This should allow us to
more robustly throw away false-negatives near the isocenter.  I also
made a few changes to the notation.  In particular, I altered our
definition of S such that it is now shifting the points in A into B,
instead of the other way around.  This allows us to keep the MRI
machine's isocenter at the origin of the final coordinate system.

Refs #3.