Fix subtle bug in feature detection
When performing the center of mass calculations within the labeled
volumes, it is probably not appropriate to use the raw data as the
"mass" array.  This is because it may have negative values and its
scaling is probably not robust.  Instead, we should use the convolved
data as the mass array.  This should result in a better result because
the values will always be positive.

Recall, one of the reasons to do the center of mass is to get sub-pixel
resoltion in the peak localization.