Split out `populate_base_content`
Typical tradeoff between DRY and KISS.  KISS won.

The `rho` that is used is different in the feature detection test from
the other two tests; also, the metrics used in feature detection are
different.  Moving forward, it seems likely other differences will show
up as well.  As a result, it seemed best to live with a bit of
duplication.