Add wrappers for file io
There were a few functions in test_utils that started to do this, but
they were not used everywhere.  Added wrappers for all file types as
well as reading and writing, and applied through out the whole project.

This is not well tested at the moment.

This refactor should allow us to move over to HDF5 at some point.