Start adding django numpy fields
There is a third party package that is similar to this, however, it
saves the data in the database---we want to have a file field that
automatically switches back and forth between numpy arrays, and saves
the files as part of the field's `pre_save` method.

I ran out of time, so the more interesting field (the file field) is
halfway finished.  For now, I am using the NumpyTextField everywhere,
however, the larger numpy arrays should be switched over to the
NumpyFileField once it is completed.