Convert to FCNN for reference.
Most likely we will not be using this because a GPU is needed to run the predictions. It is very slow on CPU. There are advantages to this approach however as we can train different types of phantoms very easily by just showing a whole bunch of examples rather than hand tuning the convolutional kernels. It will also be more resilient to slight scale and translational changes.