Update how we detect the isocenter
Previously, we had been using a hack to detect the location of the
isocenter within B's coordinate system---we used the center of mass of
all of B's fiducials.  This is bad for several reason:

- For some phantoms (e.g. the 603A) the center of mass of even the
perfect CAD points is not in the center
- If the FOV cuts off some fiducials, this could break down
- It is sensitive to noise

After some discussion with he client, we realized we can assume that:

- The isocenter is in the center of the FOV of the image
- The phantom will be centered at this location

This commit switches to the center of FOV approach.