Make "frame of reference" nullable
It should not allowed to be nullable for the MRIs, but only for the CTs.
Make the field nullable so that we can set it to an empty value.

Also delay creating the machine/sequence creation until after
validation.

And improve the error message that occurs when the points are uploaded.