Rotate and translate 604 CAD data points
Rotate by 180 degrees about the z axis, an translate by 40 along the
y-axis.  The original points that were provided to us seemed incorrect,
since it would have meant that the flat part of the 604 phantom was
pointing up in the air within the imaging bore.  Our only 604 dataset we
have now is oriented the opposite way.