Tweak celery settings
I believe all our tasks are indemopotent, so we can acknowledge them
late.  This ensures that they will be restarted if Heroku kills the dyno
while a long-running task is in progress (e.g. the overlay generation).

Also change a setting that will kill each worker after a single task is
run; this will help us control memory leaks in the short term.