Hide unnecessary admin pages
Mircea tried changing a user's password via the admin; this was a
reasonable thing to try, since it says "password", however really it is
the password hash.  To avoid this problem in the future I hid the
password field.

Also hide some models that are not really useful to have, and have
caused memory issues when Mircea tried loading them.  The memory issue
was caused because some of the models will automatically load big text
fields with ndarrays into memory.