Pull overlay interpolation code out of server
Ideally, all of heavy the processing code will be in the "process"
folder, and not in the "server" folder.

This refactor pulls the overlay interpolation code into the process
folder, and fixes the full algorithm hdats suite in the process; this
suite depended on a file that I accidentally deleted:
"process/interpolat.py".