Make the registration phase robust to large shifts
The commits were too interconnected to logically split up, so there are
a few changes here.

1. We add a factor to our objective function that prevents large
rotations from occuring

2. We remove the brute force stage; we weren't using it, although we may
need to add it back if we need to support multiple orientations for the
phantoms (e.g. a phantom can be inserted into the imaging bore front
first or back first).

See docstrings in the commit for more details about the changes.