Use a more lenient rho function during the initial convex optimization.
Before, if the initial guess had the detected points sufficiently far from the golden points, no points would be registered and the optimizer may find a local minimum that isn't aligned with the grid. This commit ensures that the initial guess is within a desired local minimum that will align the grids correctly.