Fix failing registration test
This failure was introduced by commit
72b8ad08469b7be61498c9a9315430fb79fa6218

This commit used a more lenient rho function for our first pass on the
registration, however, the first pass rho function is built using a g_cutoff of 0,
which causes the `build_rho` function to always return the
`max_match_distance`.  Since this test had a `grid_spacing` of 1, the
`max_match_distance` was negative!  By changing the `overlap_buffer`
such that it is tied to the grid_spacing, we avoid this sitution.