More gracefully handle errors that occur while parsing the DICOM files
If there is an error while saving the meta_data on the `dicom_series`,
then if a user attempts to "re-run" the scan, it was failing in an
ungraceful way because we had implicitly assumed that the meta data was
present.

This commit fixes this, and no longer assumes that the dicom_series meta
data was save correctly.

I believe the error message:

    TypeError: Cannot cast array data from dtype('float64') to
    dtype('S32') according to the rule 'safe'

is fixed by this commit.