Avoid storing voxels in the database
We had stored both the raw DICOM files AND the voxels in the database;
this was causing issues when there were very large DICOM files, so it
was simplest to simply re-extract the voxels from the DICOM files when
needed.  This has the added benefit of greatly reducing the size of the
database, which would have likely become an issue realtiavly soon (the
Heroku database was already 5Gig).