Simplify the main process scan task
It had grown to be quite large, and was hard to follow.  Split out
various steps into self-contained functions that are easier to reason
about.

Note that I tried to keep most of the high-level AlgorithmExceptions at
the top-level, since it seemed more appropriate then having them
embedded in the lower leve _save_xyz functions.