Fix bug
This bug is subtle, but could lead to very odd and hard to find
problems.

Basically, when we create a new scan or re-run an existing scan, we copy
the current golden fiducials to the scan.  This ensures that we know
which golden fiducials were set at the time when the scan was started.
We had been grabbiung the golden fiducials off of the phantom, which
would be the most "up to date" golden fiducials.

This would cause issues in a variety of side cases, for example:

1. Someone uploads a scan, but the server has a queue and it isn't
processed immediatly
2. The user changes the golden standard
3. The re-run the first scan expecting it to use the new golden standard

Before this fix, both 1 and 3 would be identical, even though the user
would expect them not to be.  After this fix, 1 will point to the
original golden points that were saved onto the scan at creation time,
and thus 1 and 3 will have different results (as expected).

I think Mircea ran into this bug once a while back, but I couldn't
figure it out at that point in time.