Optimize memory usage
1. Delete superfluous copies of the voxels
2. Perform gaussian blurring one axis at a time (since it is separable)

This second change uses significantly less memory (e.g. 1GB vs 8GB) on
one dataset I was testing.

Note that the boundary conditions are different (reflect vs constant 0);
this should reeduce the number of peaks near the edges of images due to
roll off from the zeros, but there is a slight possibility that it coudl
have issues if a phantom was adjacent to the dege of the FOV; this seems
very unlikely in practice, and probably reflect would still be better
than constant 0's even in this case.