Shorten error message text
I hit this page on my macbook, and it caused the table to wrap in an odd
way.  I manually shortend the text using the devtools, and the issue
went away.  Thus, the easiest solution was to trim the error message
text.  Since this text shouldn't be displayed except in rare cases, this
seemed like not a big issue.