Run the false positives through the FP rejector in chunks
A dataset was failing on the production server because it ran
out of memory at the false positive rejector stage.

The most obvious way to save memory in the FP rejector was to run the
points through in chunks.

The dataset that was failing had 85,000 points when starting the FP
rejector phase.  Each "window" is 15x15x15 doubles, and thus the windows
alone take:

15x15x15x85000x8 Bytes = 2.3 GBytes

of space.  I suspect Tensor flow takes quite a bit on top of this.