Handle cases where the ROI around detected peaks has a greater surface maximum intensity than the peak itself
For the 603A phantom, this can happen around the edges if the ROI includes parts of the housing.
If this occurs, we use the maximum value of the maximum intensities of the 6 ROI sides that is
less than the peak to compute the threshold value. Then, if after thresholding the ROI, there are
sections touching the edges of the ROI, we throw those parts away and take the center of mass of
what's left.