Avoid binary dilation
As described in the comment that was deleted in this commit, the binary
dilation was necessary to:

1. Expand the neighborhood we sub-voxel-zoom into
2. Merge adjacent peaks that happened to have the same maximum value

Now that we are using the COM algorithm, we no longer need the sub-voxel
zoom (1).  The sub-voxel-zoom should only ever move the peak location by a
small amount---small enough that the result of doing the COM on the
original location vs the sub-voxel-zoomed location should be identical.

We still need to handle (2).  This is done by performing an initial
binary center-of-mass on the labeled regions.