Prune datasets; make naming consistent; add notes
Ideally we wouldn't need to duplicate the phantom and the modality so
many times.  Really, the modality should be derived from the DICOM tags
like we do in production, and the phantom can be derived from the key or
the file name.

Prune out some datasets that seem redundant and add in a few new ones
that look interesting and different.