Project against invalid scan rows
We were seeing errors caused becayse we tried calling `.max` on `None`.
I looked through the database, and there were rows in `common_scan`
where:

- TP_A_S_id is NULL
- TP_B_id is NULL
- errors is NULL
- processing is False

Of all of these rows, all of them were deleted except one from
2018-10-11 (pk=1034).

It shouldn't be possible for a scan to be in this state, since if
processing is complete either there should be data (TP_A_S_id and
TP_B_id have values) or errors should be logged ("errors" is not NULL).

I looked through our tasks.py code a bit, and nothing obvious jumped out
at me.  Thus, I went ahead and added a check to our the Scan.processing
property which should fix our immediate problem

The deeper problem of how scans can be in this state is still not
solved.